requirements.txt
requests~=2.28.2
pillow~=9.2.0